+++ src/features/auth-register/ui/CreateNewPassForm/CreateNewPassForm.module.scss
