+++ src/features/auth-register/ui/CreateNewPassForm/CreateNewPassForm.tsx
