+++ src/features/auth-register/ui/ForgotPassForm/ForgotPassForm.module.scss
