+++ src/features/auth-register/ui/LoginForm/LoginForm.tsx
@@ -1,0 +1,9 @@
+// import React from 'react';
+//
+// export const LoginForm = () => {
+//   return (
+//     <div>
+//
+//     </div>
+//   );
+// };
