+++ src/features/auth-register/ui/SignUpForm/SignUpForm.module.scss
