--- src/features/auth/ui/SignTEMP/Sign.module.scss
@@ -1,34 +1,0 @@
-.formContainer {
-  display: flex;
-  gap: 1.5rem;
-  flex-direction: column;
-  align-items: center;
-  padding: 1.5rem;
-  width: 23.625rem;
-  background-color: #171717;
-  border: solid 1px #333333;
-  margin: 0 auto;
-  height: max-content;
-}
-.socialIconContainer {
-  display: flex;
-  gap: 3.75rem;
-  justify-content: center;
-}
-.signInForgotText {
-  width: 100%;
-  text-align: right;
-  margin-top: 1.5rem;
-}
-.forgotHelpText {
-  color: var(--color-light-1000);
-  margin-top: -1rem;
-}
-.helpText {
-  margin-top: -0.3rem;
-}
-.oppositeBtn {
-  border: none !important;
-  margin-top: -1rem;
-}
-
