+++ src/pages-fs/auth-register/ui/CreateNewPassword/CreateNewPassword.tsx
@@ -1,6 +1,6 @@
 import styles from './CreateNewPassword.module.scss'
 
-import CreateNewPassForm from '@/features/auth/ui/CreateNewPassForm/CreateNewPassForm'
+import CreateNewPassForm from '@/features/auth-register/ui/CreateNewPassForm/CreateNewPassForm'
 import FormContainer from '@/shared/ui/FormContainer/FormContainer'
 
 export const CreateNewPassword = () => {
