+++ src/pages-fs/auth-register/ui/ForgotPassword/ForgotPassword.tsx
@@ -1,6 +1,6 @@
 import styles from './ForgotPassword.module.scss'
 
-import ForgotPassForm from '@/features/auth/ui/ForgotPassForm/ForgotPassForm'
+import ForgotPassForm from '@/features/auth-register/ui/ForgotPassForm/ForgotPassForm'
 import FormContainer from '@/shared/ui/FormContainer/FormContainer'
 
 export const ForgotPassword = () => {
