+++ src/pages-fs/auth-register/ui/SignIn/SignIn.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import SignInForm from '@/features/auth/ui/SignInForm/SignInForm'
+import SignInForm from '@/features/auth-register/ui/SignInForm/SignInForm'
 import FormContainer from '@/shared/ui/FormContainer/FormContainer'
 
 export const SignIn = () => {
