+++ src/pages-fs/auth-register/ui/SignUp/SignUp.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import SignUpForm from '@/features/auth/ui/SignUpForm/SignUpForm'
+import SignUpForm from '@/features/auth-register/ui/SignUpForm/SignUpForm'
 import FormContainer from '@/shared/ui/FormContainer/FormContainer'
 
 export const SignUp = () => {
